Splitting file checks into two files
