+++ CIMS/model_validation/registry.py
@@ -25,7 +25,7 @@ To register a new FILE-phase error check from `file_checks.py`:
     REGISTRY.register(
         "my_new_check",
         CheckSpec(
-            fn=validate_file.my_new_check,
+            fn=file_warnings.my_new_check,
             phase=Phase.FILE,
             severity=Severity.ERROR,
             argmap={},  # validator is provided automatically
@@ -35,8 +35,8 @@ To register a new FILE-phase error check from `file_checks.py`:
 
 from dataclasses import dataclass
 from typing import Callable, Dict, Optional
-from . import file_checks as validate_file
-
+from . import file_errors
+from . import file_warnings
 
 # ---------------------------------------------------------------------------
 # Core registry tags
@@ -110,7 +110,7 @@ REGISTRY = CheckRegistry()
 # ---------------------------------------------------------------------------
 # FILE-PHASE CHECKS
 # ---------------------------------------------------------------------------
-# All of these are run by ModelValidator.validate_files() via:
+# All of these are run by ModelValidator.file_warningss() via:
 #   REGISTRY.iter(phase=Phase.FILE, severity=...)
 #
 # Errors first, then warnings.
@@ -118,109 +118,109 @@ REGISTRY = CheckRegistry()
 
 # ---- Errors ----
 REGISTRY.register("invalid_competition_type", CheckSpec(
-    fn=validate_file.invalid_competition_type,
+    fn=file_warnings.invalid_competition_type,
     phase=Phase.FILE,
     severity=Severity.ERROR,
     argmap={},
 ))
 REGISTRY.register("nodes_no_provided_service", CheckSpec(
-    fn=validate_file.nodes_no_provided_service,
+    fn=file_warnings.nodes_no_provided_service,
     phase=Phase.FILE,
     severity=Severity.ERROR,
     argmap={}
 ))
 REGISTRY.register("nodes_requesting_self", CheckSpec(
-    fn=validate_file.nodes_requesting_self,
+    fn=file_warnings.nodes_requesting_self,
     phase=Phase.FILE,
     severity=Severity.ERROR,
     argmap={}
 ))
 REGISTRY.register("supply_without_lcc_or_price", CheckSpec(
-    fn=validate_file.supply_without_lcc_or_price,
+    fn=file_warnings.supply_without_lcc_or_price,
     phase=Phase.FILE,
     severity=Severity.ERROR,
     argmap={}
 ))
 REGISTRY.register("lcc_at_tech_node", CheckSpec(
-    fn=validate_file.lcc_at_tech_node,
+    fn=file_warnings.lcc_at_tech_node,
     phase=Phase.FILE,
     severity=Severity.ERROR,
     argmap={}
 ))
 REGISTRY.register("lcc_at_tech", CheckSpec(
-    fn=validate_file.lcc_at_tech,
+    fn=file_warnings.lcc_at_tech,
     phase=Phase.FILE,
     severity=Severity.ERROR,
     argmap={}
 ))
 REGISTRY.register("nodes_with_zero_output", CheckSpec(
-    fn=validate_file.nodes_with_zero_output,
+    fn=file_warnings.nodes_with_zero_output,
     phase=Phase.FILE,
     severity=Severity.ERROR,
     argmap={}
 ))
 REGISTRY.register("undefined_nodes", CheckSpec(
-    fn=validate_file.undefined_nodes,
+    fn=file_warnings.undefined_nodes,
     phase=Phase.FILE,
     severity=Severity.ERROR,
     argmap={"providers": "providers", "requested": "requested"}
 ))
 REGISTRY.register("inconsistent_tech_refs", CheckSpec(
-    fn=validate_file.inconsistent_tech_refs,
+    fn=file_warnings.inconsistent_tech_refs,
     phase=Phase.FILE,
     severity=Severity.ERROR,
     argmap={}
 ))
 REGISTRY.register("tech_compete_nodes_no_techs", CheckSpec(
-    fn=validate_file.tech_compete_nodes_no_techs,
+    fn=file_warnings.tech_compete_nodes_no_techs,
     phase=Phase.FILE,
     severity=Severity.ERROR,
     argmap={}
 ))
 REGISTRY.register("techs_no_base_market_share", CheckSpec(
-    fn=validate_file.techs_no_base_market_share,
+    fn=file_warnings.techs_no_base_market_share,
     phase=Phase.FILE,
     severity=Severity.ERROR,
     argmap={}
 ))
 REGISTRY.register("service_req_at_tech_node", CheckSpec(
-    fn=validate_file.service_req_at_tech_node,
+    fn=file_warnings.service_req_at_tech_node,
     phase=Phase.FILE,
     severity=Severity.ERROR,
     argmap={}
 ))
 REGISTRY.register("revenue_recycling_at_techs", CheckSpec(
-    fn=validate_file.revenue_recycling_at_techs,
+    fn=file_warnings.revenue_recycling_at_techs,
     phase=Phase.FILE,
     severity=Severity.ERROR,
     argmap={}
 ))
 REGISTRY.register("both_cop_p2000_defined", CheckSpec(
-    fn=validate_file.both_cop_p2000_defined,
+    fn=file_warnings.both_cop_p2000_defined,
     phase=Phase.FILE,
     severity=Severity.ERROR,
     argmap={}
 ))
 REGISTRY.register("min_max_conflicts", CheckSpec(
-    fn=validate_file.min_max_conflicts,
+    fn=file_warnings.min_max_conflicts,
     phase=Phase.FILE,
     severity=Severity.ERROR,
     argmap={}
 ))
 REGISTRY.register("new_nodes_in_scenario", CheckSpec(
-    fn=validate_file.new_nodes_in_scenario,
+    fn=file_warnings.new_nodes_in_scenario,
     phase=Phase.FILE,
     severity=Severity.ERROR,
     argmap={}
 ))
 REGISTRY.register("new_techs_in_scenario", CheckSpec(
-    fn=validate_file.new_techs_in_scenario,
+    fn=file_warnings.new_techs_in_scenario,
     phase=Phase.FILE,
     severity=Severity.ERROR,
     argmap={}
 ))
 REGISTRY.register("base_year_market_share_not_one", CheckSpec(
-    fn=validate_file.base_year_market_share_not_one,
+    fn=file_warnings.base_year_market_share_not_one,
     phase=Phase.FILE,
     severity=Severity.ERROR,
     argmap={}
@@ -229,37 +229,37 @@ REGISTRY.register("base_year_market_share_not_one", CheckSpec(
             
 # ---- Warnings ----
 REGISTRY.register("missing_parameter_default", CheckSpec(
-    fn=validate_file.missing_parameter_default, 
+    fn=file_warnings.missing_parameter_default, 
     phase=Phase.FILE,
     severity=Severity.WARNING,
     argmap={}
 ))
 REGISTRY.register("unrequested_nodes", CheckSpec(
-    fn=validate_file.unrequested_nodes, 
+    fn=file_warnings.unrequested_nodes, 
     phase=Phase.FILE, 
     severity=Severity.WARNING, 
     argmap={"providers":"providers", "requested": "requested"}    
 ))
 REGISTRY.register("nodes_no_requested_service", CheckSpec(
-    fn = validate_file.nodes_no_requested_service, 
+    fn = file_warnings.nodes_no_requested_service, 
     phase=Phase.FILE,
     severity=Severity.WARNING,
     argmap={}
 ))
 REGISTRY.register("duplicate_service_requests", CheckSpec(
-    fn = validate_file.duplicate_service_requests, 
+    fn = file_warnings.duplicate_service_requests, 
     phase=Phase.FILE,
     severity=Severity.WARNING,
     argmap={}
 ))
 REGISTRY.register("bad_service_req", CheckSpec(
-    fn = validate_file.bad_service_req, 
+    fn = file_warnings.bad_service_req, 
     phase=Phase.FILE,
     severity=Severity.WARNING,
     argmap={}
 ))
 REGISTRY.register("zero_requested_nodes", CheckSpec(
-    fn = validate_file.zero_requested_nodes, 
+    fn = file_warnings.zero_requested_nodes, 
     phase=Phase.FILE,
     severity=Severity.WARNING,
     argmap={}
@@ -268,7 +268,7 @@ REGISTRY.register("zero_requested_nodes", CheckSpec(
 # ---------------------------------------------------------------------------
 # GRAPH-PHASE CHECKS (to be added)
 # ---------------------------------------------------------------------------
-# All of these are run by ModelValidator.validate_files() via:
+# All of these are run by ModelValidator.file_warningss() via:
 #   REGISTRY.iter(phase=Phase.GRAPH, severity=...)
 #
 # Errors first, then warnings.
